+++ goemotions/run.sh
@@ -21,4 +21,4 @@ source ./bin/activate
 pip install numpy
 pip install six
 pip install "tensorflow>=1.11.0,<2.00.0"
-pip install -r dataset_analysis/requirements.txt
+pip install -r goemotions/requirements.txt
